the initial setup of moving installations into the exhibitions fold
